+++ src/TripController.php
@@ -193,6 +193,11 @@ class TripController extends Controller {
 	}
 
 	public function deleteTrip($trip_id) {
+		//CHECK IF AGENT BOOKED TRIP VISITS
+		$agent_visits_booked = Visit::where('trip_id', $trip_id)->where('booking_method_id', 3042)->where('booking_status_id', 3061)->first();
+		if ($agent_visits_booked) {
+			return response()->json(['success' => false, 'errors' => ['Trip cannot be deleted']]);
+		}
 		$trip = Trip::where('id', $trip_id)->delete();
 		if (!$trip) {
 			return response()->json(['success' => false, 'errors' => ['Trip not found']]);
@@ -214,6 +219,12 @@ class TripController extends Controller {
 
 	public function cancelTripVisitBooking($visit_id) {
 		if ($visit_id) {
+
+			//CHECK IF AGENT BOOKED VISIT
+			$agent_visits_booked = Visit::where('booking_method_id', 3042)->where('booking_status_id', 3061)->first();
+			if ($agent_visits_booked) {
+				return response()->json(['success' => false, 'errors' => ['Visit cannot be deleted']]);
+			}
 			$visit = Visit::where('id', $visit_id)->first();
 			$visit->booking_status_id = 3062; // Booking cancelled
 			$visit->save();
